requirements.txt
accelerate==1.7.0
Adafruit-Blinka==8.58.1
adafruit-circuitpython-binascii==2.0.6
adafruit-circuitpython-busdevice==5.2.12
adafruit-circuitpython-connectionmanager==3.1.4
adafruit-circuitpython-requests==4.1.12
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.79.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.9.0
awscli==1.40.26
binho-host-adapter==0.1.6
binvox==0.1.6
binvox_rw @ file:///hpc2hdd/home/jmoon709/projects/cvpr_main/binvox_rw
bitsandbytes==0.46.0
boto3==1.38.27
botocore==1.38.27
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.0
# Editable Git install with no remote (cvpr_main==0.1)
-e /hpc2hdd/home/jmoon709/projects/cvpr_main
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docutils==0.19
ego4d==1.7.3
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
git-lfs==1.6
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_resources==6.5.2
iniconfig==2.1.0
iopath==0.1.10
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jq==1.8.0
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.84.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.0.0
pluggy==1.6.0
portalocker==3.1.1
psutil==7.0.0
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
pyftdi==0.56.0
Pygments==2.19.1
pyparsing==3.2.3
pyserial==3.5
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.7.2
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
segmentation_models_pytorch==0.5.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
sysv_ipc==1.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchaudio==2.5.1+cu121
torchmetrics==1.7.2
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
trimesh==4.6.11
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
zipp==3.22.0
